requirements.txt
opencv-python==4.4.0.44
numpy==1.19.1
scipy==1.5.0
sklearn==0.20.3
tensorboard==2.2.1
tensorboardX==2.1
torch==1.4.0
tqdm==4.46.0
Pillow==7.2.0
torchvision==0.5.0